fazle-system/social-engine/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.1
pydantic-settings==2.9.1
httpx==0.28.1
psycopg2-binary==2.9.10
cryptography==44.0.3
